Rename and move classes for better organization
