+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/dashboard/topperformers/GetSelectedRangeForTopPerformers.kt
@@ -1,8 +1,9 @@
-package com.woocommerce.android.ui.dashboard.stats
+package com.woocommerce.android.ui.dashboard.topperformers
 
 import com.woocommerce.android.AppPrefsWrapper
 import com.woocommerce.android.ui.analytics.ranges.StatsTimeRangeSelection.SelectionType
 import com.woocommerce.android.ui.dashboard.data.TopPerformersCustomDateRangeDataStore
+import com.woocommerce.android.ui.dashboard.domain.GetSelectedDateRange
 import com.woocommerce.android.util.DateUtils
 import javax.inject.Inject
